CJ 본선 신기술

+++ api-module/src/main/kotlin/com/cj/smartworker/business/day_report/port/in/StepCountUseCase.kt
@@ -3,5 +3,5 @@ package com.cj.smartworker.business.day_report.port.`in`
 import com.cj.smartworker.domain.member.entity.Member
 
 interface StepCountUseCase {
-    fun sendStepCount(member: Member, step: Long)
+    fun sendStepCount(member: Member, step: Int)
 }
